Introduce DataTableBase
