+++ packages/ra-ui-materialui/src/list/datatable/context/index.ts
@@ -3,6 +3,7 @@ export * from './DataTableColumnRankContext';
 export * from './DataTableConfigContext';
 export * from './DataTableDataContext';
 export * from './DataTableRenderContext';
+export * from './DataTableRowSxContext';
 export * from './DataTableSelectedIdsContext';
 export * from './DataTableSortContext';
 export * from './DataTableStoreContext';
